changed update_progress prototype to an int 0<percent<100
git-svn-id: svn://svn.code.sf.net/p/gphoto/code/trunk@962 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ gphoto/minolta/dimage_v.h
@@ -58,7 +58,7 @@ typedef struct _dimage_v_buffer
 /* GPhoto globals */
 extern char	serial_port[20];
 extern void update_status(char *newStatus);
-extern void update_progress(float percentage);
+extern void update_progress(int percentage);
 extern void error_dialog(char *Error);
 
 /* Globals for my own use. */
